region-service/scripts/requirements.txt
geopandas
sqlalchemy
shapely
psycopg2-binary
python-dotenv